chatsuffix > customchat
